feat: Add support for message files in modal store
The `messageFile` type has been added to the `ModalType` enum in the `use-modal-store.ts` file. This allows for the inclusion of message files in the modal data, providing support for sending and displaying files in messages. This feature enhances the user experience by enabling the attachment and sharing of files within the application.